refactor: split cooperativeAssaultValidators into focused modules
Split the monolithic cooperativeAssaultValidators.ts (528 lines) into a
modular subdirectory structure following the same pattern as combatValidators:

- helpers.ts: Shared functions (findCityHex, isAdjacentToCity, hasNonWoundCard)
- proposeValidators.ts: 11 validators for PROPOSE_COOPERATIVE_ASSAULT_ACTION
- responseValidators.ts: 3 validators for RESPOND_TO_COOPERATIVE_PROPOSAL_ACTION
- cancelValidators.ts: 2 validators for CANCEL_COOPERATIVE_PROPOSAL_ACTION
- index.ts: Re-exports all validators

This improves maintainability by grouping validators by their action type
and extracting shared helper functions.
